Move end points to controllers
